pyproject.toml
[project]
name = "pta-ciencia-de-dados-equipe-julia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=2.3.2",
    "aiosqlite>=0.21.0",
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "pandas>=2.3.3",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
]
